sci-misc/praat: Drop virtual/x11 dependency.
This wasn't problematic in practice because it was the second branch of a || ()
clause, but pkgcheck complained. virtual/x11 hasn't been in the tree since the last
in-tree reverse dependencies were ported to depend directly on the modular-X components
they need, which predates the conversion of the main ::gentoo Portage tree to git.