app-text/FanFicFare: Drop legacy /usr/bin/ffd symlink
This package used to be called 'fanficownloader', so it (either by default or as I packaged it, to save
typing) had a way of calling it via 'ffd'. So when the package was renamed I added a legacy symlink to
the packaging so I wouldn't be constantly typing the old name and getting 'command not found'. But by
now my muscle memory is well-trained to type 'fff' instead, so I'm dropping the 'ffd' symlink.